requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
requests==2.31.0
langchain==0.1.0
langchain-google-genai==0.0.6
langchain-community==0.0.10
faiss-cpu==1.7.4
sentence-transformers==2.2.2
python-multipart==0.0.6
PyPDF2==3.0.1
python-docx==1.1.0
python-dotenv==1.0.0
google-generativeai==0.7.2
